requirements.txt
flask
opencv-python-headless
numpy
pillow
gunicorn

